任务的删除与更新的相关逻辑判断

+++ src/main/java/com/program/itta/service/impl/TaskServiceImpl.java
@@ -1,6 +1,7 @@
 package com.program.itta.service.impl;
 
 import com.program.itta.common.exception.task.TaskNameExistsException;
+import com.program.itta.common.exception.task.TaskNotExistsException;
 import com.program.itta.domain.entity.Task;
 import com.program.itta.mapper.TaskMapper;
 import com.program.itta.service.TaskService;
@@ -9,6 +10,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.Date;
 import java.util.List;
 
 /**
@@ -28,11 +30,11 @@ public class TaskServiceImpl implements TaskService {
     @Override
     public Boolean addTask(Task task) {
         Boolean judgeTaskName = judgeTaskName(task);
-        if(judgeTaskName){
+        if (judgeTaskName) {
             throw new TaskNameExistsException("该任务名称已存在于此项目中");
         }
         int insert = taskMapper.insert(task);
-        if (insert!=0){
+        if (insert != 0) {
             return true;
         }
         return false;
@@ -40,9 +42,12 @@ public class TaskServiceImpl implements TaskService {
 
     @Override
     public Boolean deleteTask(Task task) {
+        Boolean judgeTaskExists = judgeTaskExists(task);
+        if (!judgeTaskExists){
+            throw new TaskNotExistsException("该任务不存在，任务id查找为空");
+        }
         int delete = taskMapper.deleteByPrimaryKey(task.getId());
-        // TODO 中间表的处理
-        if (delete!=0){
+        if (delete != 0) {
             return true;
         }
         return false;
@@ -50,8 +55,13 @@ public class TaskServiceImpl implements TaskService {
 
     @Override
     public Boolean updateTask(Task task) {
+        Boolean judgeTaskExists = judgeTaskExists(task);
+        if (!judgeTaskExists){
+            throw new TaskNotExistsException("该任务不存在，任务id查找为空");
+        }
+        task.setUpdateTime(new Date());
         int update = taskMapper.updateByPrimaryKey(task);
-        if (update!=0){
+        if (update != 0) {
             return true;
         }
         return false;
@@ -62,12 +72,20 @@ public class TaskServiceImpl implements TaskService {
         return null;
     }
 
-    private Boolean judgeTaskName(Task task){
+    private Boolean judgeTaskName(Task task) {
         List<Task> taskList = taskMapper.selectAllByItemId(task.getItemId());
         for (Task task1 : taskList) {
-            if (task.getName().equals(task1.getName())){
+            if (task.getName().equals(task1.getName())) {
                 return true;
             }
+        }
+        return false;
+    }
+
+    private Boolean judgeTaskExists(Task task) {
+        Task select = taskMapper.selectByPrimaryKey(task.getId());
+        if (select != null) {
+            return true;
         }
         return false;
     }
